split test_dss into fakers production and service test
