+++ conda/meta.yaml
@@ -36,7 +36,7 @@ test:
     - test/**
   commands:
     - cd test
-    - pytest -v --ignore=test_weatherdata.py
+    - pytest -v --ignore=test_weatherdata.py --ignore=test_dss.py
 about:
   home: {{ data.get('url') }}
   license: GPL-v3
