+++ test/test_dss.py
@@ -8,6 +8,7 @@ link = ipm.get_dss('LINK')
 onthefly = ipm.get_dss('ONTHEFLY')
 
 link_dss_models = sum([[(d, m) for m in v['models']] for d, v in link.items()], [])
+onthefly_dss_models = sum([[(d, m) for m in v['models']] for d, v in onthefly.items()], [])
 
 
 @pytest.mark.parametrize('dss,model', link_dss_models)
@@ -18,43 +19,27 @@ def test_dss_link(dss, model):
     if m['execution']['endpoint'] == '':
         assert 'description_URL' in m
     assert ipm_fakers.input_data(m) is None
-    res = ipm.run_model(m)
-    assert isinstance(res, str)
+    try:
+        res = ipm.run_model(m)
+        assert isinstance(res, str)
+    except:
+        raise
+    else:
+        print ('ok')
 
 
-onthefly_dss_models = sum([[(d, m) for m in v['models']] for d, v in onthefly.items()], [])
-noweather_nofield = [(d, m) for d, m in onthefly_dss_models if
-                                         onthefly[d]['models'][m]['input'] is None]
-input_not_none = [(d, m) for d, m in onthefly_dss_models if
-                                         onthefly[d]['models'][m]['input'] is not None]
-weather_nofield = [(d, m) for d, m in input_not_none if
-                   (onthefly[d]['models'][m]['input']['weather_parameters'] is not None)
-                   & (onthefly[d]['models'][m]['input']['field_observation'] is None)]
-field = [(d, m) for d, m in input_not_none if
-                   onthefly[d]['models'][m]['input']['field_observation'] is not None]
-
-@pytest.mark.parametrize('dss,model', noweather_nofield)
-def test_faker_dss_onthefly_noweather_nofield(dss, model):
+@pytest.mark.parametrize('dss,model', onthefly_dss_models)
+def test_dss_onthefly(dss, model):
     m = onthefly[dss]['models'][model]
     assert m['execution']['type'] == 'ONTHEFLY'
     assert 'endpoint' in m['execution']
     assert len(m['execution']['endpoint']) > 0
-    assert 'input_schema' in m['execution']
-    assert len(m['execution']['input_schema']) > 0
-    fake = ipm_fakers.input_data(m)
-    assert isinstance(fake, dict)
-
-@pytest.mark.parametrize('dss,model', weather_nofield)
-def test_faker_dss_onthefly_weather_nofield(dss, model):
-    m = onthefly[dss]['models'][model]
-    assert m['execution']['type'] == 'ONTHEFLY'
     assert 'endpoint' in m['execution']
-    assert len(m['execution']['endpoint']) > 0
-    assert 'input_schema' in m['execution']
-    assert len(m['execution']['input_schema']) > 0
-    assert 'weatherData' in m['execution']['input_schema']['properties']
     fake = ipm_fakers.input_data(m)
-    assert isinstance(fake, dict)
-    assert 'weatherData' in fake
-
-
+    try:
+        res = ipm.run_model(m, input_data=fake)
+        assert isinstance(res, dict), res
+    except:
+        raise
+    else:
+        'ok'
